+++ noodles-sam/src/alignment/io.rs
@@ -1,0 +1,5 @@
+//! Alignment record I/O.
+
+mod reader;
+
+pub use self::reader::Reader;